backend/AuthenticationService/requirements.txt
django==4.2.26
psycopg2-binary==2.9.5
wheel==0.38.4
djangorestframework==3.14.0
django[argon2]
django[bcrypt]
python-dotenv==0.21.0
Celery==5.2.7
redis==4.4.2
Celery[redis]
django_redis==5.2.0
PyJWT==2.6.0
gunicorn==20.1.0
django-cacheops==6.1
drf-orjson-renderer==1.7.1
django-rest-swagger==2.2.0
python-dotenv==0.21.0
drf_yasg==1.21.6
djangorestframework-simplejwt==5.5.1
